feat : 채팅 dto 수정

+++ src/main/java/com/multi/mlpenterpriseapprovalsystem/chat/dto/ResChatRoomListDto.java
@@ -24,24 +24,22 @@ public class ResChatRoomListDto {
     private String roomName;
     private String lastMessage;
     private LocalDateTime lastMessageAt;
-    private Long unreadCount;
+    private int unreadCount;
 
     public static ResChatRoomListDto from(ChatRoom room, String myEmpId) {
 
-        Long unreadCount = room.getMembers().stream()
+        int unreadCount = room.getMembers().stream()
                 .filter(m -> m.getEmployee().getEmpId().equals(myEmpId))
                 .findFirst()
                 .map(ChatRoomMember::getUnreadCount)
-                .orElse(0L);
+                .orElse(0);
 
-        // 나 제외 멤버 이름 목록
         List<String> otherNames = room.getMembers().stream()
                 .map(ChatRoomMember::getEmployee)
                 .filter(emp -> !emp.getEmpId().equals(myEmpId))
                 .map(Employee::getEmpName)
                 .toList();
 
-        // 그룹방이고 roomName이 있으면 그걸 우선 사용 (사용자 지정 방이름 보호)
         String displayName;
         if (room.getRoomType() == RoomType.GROUP
                 && room.getRoomName() != null
